+++ Kmd.Momentum.Mea.Integrations.Tests/Kmd.Momentum.Mea.Integrations.Tests.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.2" />
+    <PackageReference Include="xunit" Version="2.4.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\src\Kmd.Momentum.Mea.Api\Kmd.Momentum.Mea.Api.csproj" />
+  </ItemGroup>
+
+</Project>
