--- test/Kmd.Momentum.Mea.Api.Tests/FabricOfSpaceTimeTests.cs
@@ -1,14 +1,0 @@
-using FluentAssertions;
-using Xunit;
-
-namespace Kmd.Momentum.Mea.Api.Tests
-{
-    public class FabricOfSpaceTimeTests
-    {
-        [Fact]
-        public void IsCorrectUniverse()
-        {
-            true.Should().BeTrue(because: "In this universe, 'true' definitely is 'true'");
-        }
-    }
-}
